Change version to be hard-coded.
The version, a.k.a hjsv (Hotjar Snippet Version) is used to describe the "version" of the snippet code, and is not intended to be changed.
The only time it should be changed is when the actual snippet code itself changes.
Setting it incorrectly will cause Hotjar to believe you're using a different snippet version than you are.